pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "MolecularDiffusion"
version = "0.1.0"
description = "A molecular diffusion framework for machine learning applications"
readme = "README.rst"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Thanapat Worakul", email = "thanapat.worakul@epfl.ch"},
]
maintainers = [
  { name = "Thanapat Worakul", email = "thanapat.worakul@epfl.ch" },
]

dependencies = [
    "click>=8.0.0",
    "fire",
    "decorator",
    "numpy==1.26.4",
    "scipy",
    "rdkit-pypi",
    "posebusters",
    "networkx",
    "matplotlib",
    "pandas",
    "scikit-learn",
    "tqdm",
    "pyyaml",
    "omegaconf",
    "ase",
    "morfeus-ml",
    "wandb",
    "hydra-colorlog",
    "hydra-core>=1.3.0",
    "rootutils",
    "psutil",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.8.0",
]

[project.urls]
Homepage = "https://github.com/pregHosh/MolecularDiffusion"
Repository = "https://github.com/pregHosh/MolecularDiffusion"
Documentation = "https://moleculardiffusion.readthedocs.io"
"Bug Tracker" = "https://github.com/pregHosh/MolecularDiffusion/issues"

[project.scripts]
MolCraftDiff = "MolecularDiffusion.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"MolecularDiffusion" = [
    "*.py", "*.txt", "*.rst", "*.md",
    "data/**/*.json",
    "configs/**/*.yaml",
]

[tool.setuptools_scm]
write_to = "src/MolecularDiffusion/_version.py"

[tool.ruff]
line-length = 79
target-version = "py310"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN101", "ANN401", "COM812", "ISC001","ANN201"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "D105",
  "D106",
  "D107",
  "S101",
  "INP001",
  "ANN101",
  "ANN201",
]
"examples/*" = [
  "INP001",
]
"docs/source/conf.py" = ["D100", "INP001"]

[tool.mypy]
show_error_codes = true
implicit_optional = false
warn_no_return = true
strict_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_unreachable = true
disallow_any_generics = true

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true
